Create  Check if Possible to survive.java
You are a poor citizen of an island that has only one shop for groceries, which is open six days a week and is closed on Sundays. The first day of the week is Monday. You can buy ‘n’ units of food from a day’s work. Now each day, you need ‘x’ units of food to survive. You are making a boat to escape the island, which will require ‘d’ days to be made. You need to survive at least ‘d’ days on the island. You are given three integers n, x, and d. Your task is to determine if you can escape the island